Time: 0 ms (100.00%), Space: 7.1 MB (84.06%) - LeetHub

+++ 1491-average-salary-excluding-the-minimum-and-maximum-salary/1491-average-salary-excluding-the-minimum-and-maximum-salary.cpp
@@ -20,7 +20,7 @@ public:
         sum = sum - maxi - mini;
         
         double ans = 0;
-        ans = (double)(sum * 1.0) / (n-2);
+        ans = (sum * 1.0) / (n-2);
         return ans;
     }
 };
